refactor: extract Clock component from App to separate file
